Here is where MergeAndSplit ends - deciding not to do it for now
MergeAndSplit will require knowing exactly which screens players are entering and exiting, which VisualNotifier can't do. And since there are 4 separate screens, this will be a real hassle to implement what will likely be a janky mode of camera manipulation...
Like visualizing it, I'm not sure how the merge will look, and it is kinda unfair to give one player half the screen.